+++ tilematrix/__init__.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 from .tilematrix import (
+    Tile,
     TilePyramid,
     MetaTilePyramid
 )
@@ -14,5 +15,6 @@ from .io import (
     write_raster_window,
     read_vector_window,
     file_bbox,
-    get_best_zoom_level
+    get_best_zoom_level,
+    RESAMPLING_METHODS
 )