+++ tilematrix/io.py
@@ -22,9 +22,18 @@ import pyproj
 
 from tilematrix import *
 
+RESAMPLING_METHODS = {
+    "nearest": RESAMPLING.nearest,
+    "bilinear": RESAMPLING.bilinear,
+    "cubic": RESAMPLING.cubic,
+    "cubic_spline": RESAMPLING.cubic_spline,
+    "lanczos": RESAMPLING.lanczos,
+    "average": RESAMPLING.average,
+    "mode": RESAMPLING.mode
+}
+
 def read_vector_window(
     input_file,
-    tile_pyramid,
     tile,
     pixelbuffer=0
     ):
@@ -32,12 +41,6 @@ def read_vector_window(
     Reads an input vector dataset with fiona using the tile bounding box as
     filter and clipping geometry. Returns a list of GeoJSON like features.
     """
-
-    try:
-        assert (isinstance(tile_pyramid, TilePyramid) or
-            isinstance(tile_pyramid, MetaTilePyramid))
-    except:
-        raise ValueError("no valid tile matrix given.")
 
     try:
         assert os.path.isfile(input_file)
@@ -58,126 +61,56 @@ def read_vector_window(
     # TODO reproject tile_bounds to input file crs and reproject output clip to
     # tile_pyramid crs
 
-    tile_bounds = tile_pyramid.tile_bounds(*tile, pixelbuffer=pixelbuffer)
-    tile_bbox = tile_pyramid.tile_bbox(*tile, pixelbuffer=pixelbuffer)
-
     with fiona.open(input_file, 'r') as vector:
-        features_clipped = [
-            {
+        for feature in vector.filter(
+            bbox=tile.bounds(pixelbuffer=pixelbuffer)
+        ):
+            geom = tile.bbox(pixelbuffer=pixelbuffer).intersection(
+                shape(feature['geometry'])
+            )
+            feature = {
                 'properties': feature['properties'],
-                'geometry': mapping(
-                    tile_bbox.intersection(shape(feature['geometry']))
-                )
+                'geometry': mapping(geom)
             }
-            for feature in vector.filter(bbox=tile_bounds)
-        ]
 
-    return features_clipped
-
+            yield feature
 
 def read_raster_window(
     input_file,
-    tile_pyramid,
     tile,
+    indexes=None,
     pixelbuffer=0,
     resampling="nearest"
     ):
     """
-    Reads an input raster dataset with rasterio and resamples array to target
-    tile. Returns a tuple of (metadata, data), where metadata is a rasterio
-    meta dictionary and data a multidimensional numpy array.
+    Generates numpy arrays from input raster.
     """
-
-    try:
-        assert (isinstance(tile_pyramid, TilePyramid) or
-            isinstance(tile_pyramid, MetaTilePyramid))
-    except:
-        raise ValueError("no valid tile matrix given.")
-
     try:
         assert os.path.isfile(input_file)
     except:
-        raise IOError("input file does not exist: %s" % input_file)
-
-    try:
-        assert pixelbuffer >= 0
-    except:
-        raise ValueError("pixelbuffer must be 0 or greater")
-
-    try:
-        assert isinstance(pixelbuffer, int)
-    except:
-        raise ValueError("pixelbuffer must be an integer")
-
-    resampling_methods = {
-        "nearest": RESAMPLING.nearest,
-        "bilinear": RESAMPLING.bilinear,
-        "cubic": RESAMPLING.cubic,
-        "cubic_spline": RESAMPLING.cubic_spline,
-        "lanczos": RESAMPLING.lanczos,
-        "average": RESAMPLING.average,
-        "mode": RESAMPLING.mode
-    }
-
-    try:
-        assert resampling in resampling_methods
-    except:
-        raise ValueError("resampling method %s not found." % resampling)
-
-    zoom, row, col = tile
-
-    try:
-        src_bbox = file_bbox(input_file, tile_pyramid)
-    except:
-        raise
-    dst_tile_size = tile_pyramid.tile_size + 2 * pixelbuffer
-    dst_shape = (dst_tile_size, dst_tile_size)
-    tile_geom = tile_pyramid.tile_bbox(zoom, row, col, pixelbuffer)
+        raise IOError("input file not found %s" % input_file)
     with rasterio.open(input_file, "r") as src:
-        out_meta = src.meta
+
+        if indexes:
+            if isinstance(indexes, list):
+                band_indexes = indexes
+            else:
+                band_indexes = [indexes]
+        else:
+            band_indexes = src.indexes
+
+        # Reproject tile bounds to source file SRS.
+        src_left, src_bottom, src_right, src_top = transform_bounds(
+            tile.crs,
+            src.crs,
+            *tile.bounds(pixelbuffer=pixelbuffer),
+            densify_pts=21
+            )
+
         nodataval = src.nodata
         # Quick fix because None nodata is not allowed.
         if not nodataval:
             nodataval = 0
-
-        # Return array filled with NAN values if tile does not intersect with
-        # input raster.
-        try:
-            assert tile_geom.intersects(src_bbox)
-        except:
-            out_data = ()
-            for i, dtype in zip(src.indexes, src.dtypes):
-                zeros = np.zeros(shape=(dst_shape), dtype=dtype)
-                out_band = ma.masked_array(
-                    zeros,
-                    mask=True
-                )
-                out_data += (out_band,)
-            return out_meta, out_data
-
-        # Get tile bounds including pixel buffer.
-        tile_left, tile_bottom, tile_right, tile_top = tile_pyramid.tile_bounds(
-            zoom,
-            row,
-            col,
-            pixelbuffer
-            )
-
-        # Create tile affine
-        px_size = tile_pyramid.pixel_x_size(zoom)
-        tile_geotransform = (tile_left, px_size, 0.0, tile_top, 0.0, -px_size)
-        tile_affine = Affine.from_gdal(*tile_geotransform)
-
-        # Reproject tile bounds to source file SRS.
-        src_left, src_bottom, src_right, src_top = transform_bounds(
-            tile_pyramid.crs,
-            src.crs,
-            tile_left,
-            tile_bottom,
-            tile_right,
-            tile_top,
-            densify_pts=21
-            )
 
         minrow, mincol = src.index(src_left, src_top)
         maxrow, maxcol = src.index(src_right, src_bottom)
@@ -191,50 +124,38 @@ def read_raster_window(
         window_affine = src.affine * window_vector_affine
 
         # Finally read data per band and store it in tuple.
-        out_data = ()
-        for index, dtype in zip(
-            src.indexes,
-            src.dtypes
-            ):
-            src_band_data = _read_band_to_tile(
-                index,
-                src,
-                window,
-                nodataval
+        bands = (
+            src.read(index, window=window, masked=True, boundless=True)
+            for index in band_indexes
             )
-            dst_band_data = np.zeros(shape=(dst_shape), dtype=dtype)
-            dst_band_data[:] = nodataval
-            try:
-                reproject(
-                    src_band_data,
-                    dst_band_data,
-                    src_transform=window_affine,
-                    src_crs=src.crs,
-                    src_nodata=nodataval,
-                    dst_transform=tile_affine,
-                    dst_crs=tile_pyramid.crs,
-                    dst_nodata=nodataval,
-                    resampling=resampling_methods[resampling]
-                )
-            except:
-                raise
-            dst_band_data = ma.masked_equal(dst_band_data, nodataval)
-            out_data = out_data + (dst_band_data, )
-
-        # Create output metadata
-        out_meta = src.meta
-        out_meta['nodata'] = nodataval
-        out_meta['affine'] = tile_affine
-        out_meta['transform'] = tile_geotransform
-        out_meta['height'] = dst_tile_size
-        out_meta['width'] = dst_tile_size
-
-    return out_meta, out_data
+        for index in band_indexes:
+            dst_band = np.ma.zeros(
+                (tile.shape(pixelbuffer=pixelbuffer)),
+                src.dtypes[index-1]
+            )
+            dst_band[:] = nodataval
+            reproject(
+                next(bands),
+                dst_band,
+                src_transform=window_affine,
+                src_crs=src.crs,
+                src_nodata=nodataval,
+                dst_transform=tile.affine(pixelbuffer=pixelbuffer),
+                dst_crs=tile.crs,
+                dst_nodata=nodataval,
+                resampling=RESAMPLING_METHODS[resampling]
+            )
+            dst_band = ma.masked_equal(dst_band, nodataval)
+            dst_band = ma.masked_array(
+                dst_band,
+                mask=ma.fix_invalid(dst_band, fill_value=0).mask
+            )
+            dst_band.harden_mask()
+            yield dst_band
 
 
 def write_raster_window(
     output_file,
-    tile_pyramid,
     tile,
     metadata,
     bands,
@@ -242,12 +163,6 @@ def write_raster_window(
     """
     Writes numpy array into a TilePyramid tile.
     """
-    try:
-        assert (isinstance(tile_pyramid, TilePyramid) or
-            isinstance(tile_pyramid, MetaTilePyramid))
-    except:
-        raise ValueError("no valid tile matrix given")
-
     try:
         assert pixelbuffer >= 0
     except:
@@ -258,83 +173,60 @@ def write_raster_window(
     except:
         raise ValueError("pixelbuffer must be an integer")
 
-    zoom, row, col = tile
-
     # get write window bounds (i.e. tile bounds plus pixelbuffer) in affine
-    dst_left, dst_bottom, dst_right, dst_top = tile_pyramid.tile_bounds(zoom,
-        row, col, pixelbuffer)
+    dst_left, dst_bottom, dst_right, dst_top = tile.bounds(pixelbuffer)
 
-    dst_width = tile_pyramid.tile_size + (pixelbuffer * 2)
-    dst_height = tile_pyramid.tile_size + (pixelbuffer * 2)
-    pixel_x_size = tile_pyramid.pixel_x_size(zoom)
-    pixel_y_size = tile_pyramid.pixel_y_size(zoom)
-    dst_affine = calculate_default_transform(
-        tile_pyramid.crs,
-        tile_pyramid.crs,
-        dst_width,
-        dst_height,
-        dst_left,
-        dst_bottom,
-        dst_right,
-        dst_top,
-        resolution=(
-            tile_pyramid.pixel_x_size(zoom),
-            tile_pyramid.pixel_y_size(zoom)
-        ))[0]
+    dst_width, dst_height = tile.shape(pixelbuffer)
+    pixel_x_size = tile.pixel_x_size
+    pixel_y_size = tile.pixel_y_size
+    dst_affine = tile.affine(pixelbuffer)
 
-    # convert to pixel coordinates
-    affine = metadata["affine"]
-    input_left, input_top = affine * (0, 0)
-    input_right, input_bottom = affine * (metadata["width"], metadata["height"])
-    ul = input_left, input_top
-    ur = input_right, input_top
-    lr = input_right, input_bottom
-    ll = input_left, input_bottom
-    px_left = int(round(((dst_left - input_left) / pixel_x_size), 0))
-    px_bottom = int(round(((input_top - dst_bottom) / pixel_y_size), 0))
-    px_right = int(round(((dst_right - input_left) / pixel_x_size), 0))
-    px_top = int(round(((input_top - dst_top) / pixel_y_size), 0))
+    # determine pixelbuffer from shape and determine pixel window
+    src_pixelbuffer = (bands[0].shape[0] - tile.width) / 2
+    px_top, px_left = src_pixelbuffer, src_pixelbuffer
+    other_bound = src_pixelbuffer + tile.width
+    px_bottom, px_right = other_bound, other_bound
+
     window = (px_top, px_bottom), (px_left, px_right)
 
     dst_bands = []
 
-    if tile_pyramid.format.name == "PNG_hillshade":
+    if tile.tile_pyramid.format.name == "PNG_hillshade":
         zeros = np.zeros(bands[0][px_top:px_bottom, px_left:px_right].shape)
         for band in range(1,4):
+            band = np.clip(band, 0, 255)
             dst_bands.append(zeros)
 
     for band in bands:
         dst_bands.append(band[px_top:px_bottom, px_left:px_right])
 
-    if tile_pyramid.format.name == "PNG":
-        mask = ma.getmask(bands[0], )
-        nodata_alpha = np.zeros(bands[0].shape)
-        nodata_alpha[:] = 255
-        nodata_alpha[mask] = 0
-        dst_bands.append(nodata_alpha[px_top:px_bottom, px_left:px_right])
+    bandcount = tile.tile_pyramid.format.profile["count"]
 
+    if tile.tile_pyramid.format.name == "PNG":
+        for band in bands:
+            band = np.clip(band, 0, 255)
+        if tile.tile_pyramid.format.profile["nodata"]:
+            nodata_alpha = np.zeros(bands[0].shape)
+            nodata_alpha[:] = 255
+            nodata_alpha[bands[0].mask] = 0
+            dst_bands.append(nodata_alpha[px_top:px_bottom, px_left:px_right])
+            bandcount += 1
 
     # write to output file
-    dst_metadata = deepcopy(tile_pyramid.format.profile)
+    dst_metadata = deepcopy(tile.tile_pyramid.format.profile)
     dst_metadata.pop("transform", None)
-    dst_metadata["nodata"] = metadata["nodata"]
-    dst_metadata["crs"] = tile_pyramid.crs['init']
+    dst_metadata["crs"] = tile.crs['init']
     dst_metadata["width"] = dst_width
     dst_metadata["height"] = dst_height
     dst_metadata["affine"] = dst_affine
-    dst_metadata["affine"] = dst_affine
-    dst_metadata["count"] = len(dst_bands)
-    dst_metadata["dtype"] = dst_bands[0].dtype.name
-    try:
-        dst_metadata.update(compress=metadata["compress"])
-    except:
-        pass
-    try:
-        dst_metadata.update(predictor=metadata["predictor"])
-    except:
-        pass
-    if tile_pyramid.format.name in ("PNG", "PNG_hillshade"):
-        dst_metadata.update(dtype='uint8')
+
+    if tile.tile_pyramid.format.name in ("PNG", "PNG_hillshade"):
+        dst_metadata.update(
+            dtype='uint8',
+            count=bandcount
+        )
+
+    assert len(dst_bands) == dst_metadata["count"]
     with rasterio.open(output_file, 'w', **dst_metadata) as dst:
         for band, data in enumerate(dst_bands):
             data = np.ma.filled(data, dst_metadata["nodata"])
@@ -357,17 +249,18 @@ def file_bbox(
     """
     out_crs = tile_pyramid.crs
     # Read raster data with rasterio, vector data with fiona.
-    try:
+    extension = os.path.splitext(input_file)[1][1:]
+    if extension in ["shp", "geojson"]:
+        with fiona.open(input_file) as inp:
+            inp_crs = inp.crs
+            left, bottom, right, top = inp.bounds
+    else:
         with rasterio.open(input_file) as inp:
             inp_crs = inp.crs
             left = inp.bounds.left
             bottom = inp.bounds.bottom
             right = inp.bounds.right
             top = inp.bounds.top
-    except IOError:
-        with fiona.open(input_file) as inp:
-            inp_crs = inp.srs
-            left, bottom, right, top = inp.bounds
 
     # Create bounding box polygon.
     tl = [left, top]
@@ -449,93 +342,7 @@ def get_best_zoom_level(input_file, tile_pyramid_type):
     )
 
     for zoom in range(0, 25):
-        print tile_pyramid.pixel_x_size(zoom), avg_resolution
-        if (tile_pyramid.pixel_x_size(zoom) < avg_resolution):
+        if (tile_pyramid.pixel_x_size(zoom) <= avg_resolution):
             return zoom-1
 
     raise ValueError("no fitting zoom level found")
-
-
-def _read_band_to_tile(
-    index,
-    src,
-    window,
-    nodataval
-    ):
-    """
-    Reads window of a raster and if window is outside of input raster bounds,
-    fills these values with numpy nan. Returns a metadata and data tuple.
-    """
-
-    (minrow, maxrow), (mincol, maxcol) = window
-    window_offset_row = minrow
-    window_offset_col = mincol
-    minrow, minrow_offset = _clean_coord_offset(
-        minrow,
-        src.shape[0]
-        )
-    maxrow, maxrow_offset = _clean_coord_offset(
-        maxrow,
-        src.shape[0]
-        )
-    mincol, mincol_offset = _clean_coord_offset(
-        mincol,
-        src.shape[1]
-        )
-    maxcol, maxcol_offset = _clean_coord_offset(
-        maxcol,
-        src.shape[1]
-        )
-    rows = (minrow, maxrow)
-    cols = (mincol, maxcol)
-
-    window_data = src.read(index, window=(rows, cols), masked=True)
-    if minrow_offset:
-        nullarray = np.zeros(
-            (minrow_offset, window_data.shape[1]),
-            dtype=window_data.dtype
-            )
-        nullarray[:] = nodataval
-        newarray = np.concatenate((nullarray, window_data), axis=0)
-        window_data = newarray
-    if maxrow_offset:
-        nullarray = np.zeros(
-            (maxrow_offset, window_data.shape[1]),
-            dtype=window_data.dtype
-            )
-        nullarray[:] = nodataval
-        newarray = np.concatenate((window_data, nullarray), axis=0)
-        window_data = newarray
-    if mincol_offset:
-        nullarray = np.zeros(
-            (window_data.shape[0], mincol_offset),
-            dtype=window_data.dtype
-            )
-        nullarray[:] = nodataval
-        newarray = np.concatenate((nullarray, window_data), axis=1)
-        window_data = newarray
-    if maxcol_offset:
-        nullarray = np.zeros(
-            (window_data.shape[0], maxcol_offset),
-            dtype=window_data.dtype
-            )
-        nullarray[:] = nodataval
-        newarray = np.concatenate((window_data, nullarray), axis=1)
-        window_data = newarray
-
-    return window_data
-
-
-def _clean_coord_offset(px_coord, maximum):
-    """
-    Crops pixel coordinate to 0 and maximum (array.shape) if necessary. Returns
-    new pixel value and an offset if necessary.
-    """
-    offset = None
-    if px_coord < 0:
-        offset = -px_coord
-        px_coord = 0
-    if px_coord > maximum:
-        offset = px_coord - maximum
-        px_coord = maximum
-    return px_coord, offset